+++ actor/pyproject.toml
@@ -4,9 +4,10 @@ version = "0.1.0"
 description = ""
 authors = ["Your Name <you@example.com>"]
 readme = "README.md"
+packages = [{include = "actor"}]
 
 [tool.poetry.scripts]
-main = "actor.main:main"
+main = "actor.main:run"
 
 [tool.poetry.dependencies]
 python = "^3.11"
